Separate E2EE witness protocol
Move witness batching, cursor management, repair, validation, and cancellation tests into a focused module while preserving the db-app API and runtime behavior.
